chore(patentes): move PatentesService to shared/services and update imports
